requirements.txt
Flask
wikipedia-api
pytesseract
Pillow
gunicorn
